clean up and separate import/export code, allow exporting categories in JSON
