+++ ui/core/components/sim_header.tsx
@@ -4,8 +4,8 @@ import { ref } from 'tsx-vanilla';
 import { SimUI } from '../sim_ui';
 import { isLocal } from '../utils';
 import { Component } from './component';
-import { Exporter } from './exporters';
-import { Importer } from './importers';
+import { Exporter } from './exporter';
+import { Importer } from './importer';
 import { SettingsMenu } from './settings_menu';
 import { SimTab } from './sim_tab';
 import { SocialLinks } from './social_links';
