+++ ui/raid/settings_tab.ts
@@ -1,19 +1,19 @@
-import { ContentBlock } from "../core/components/content_block";
-import { EncounterPicker } from "../core/components/encounter_picker";
-import { IconPicker } from "../core/components/icon_picker";
-import { SavedDataManager } from "../core/components/saved_data_manager";
-import { SimTab } from "../core/components/sim_tab";
-import * as Tooltips from "../core/constants/tooltips.js";
-import { Encounter } from "../core/encounter";
-import { RaidBuffs } from "../core/proto/common";
-import { SavedEncounter } from "../core/proto/ui";
-import { ActionId } from "../core/proto_utils/action_id";
-import { Raid } from "../core/raid";
-import { EventID } from "../core/typed_event";
-import { AssignmentsPicker } from "./assignments_picker";
-import { BlessingsPicker } from "./blessings_picker";
-import { RaidSimUI } from "./raid_sim_ui";
-import { TanksPicker } from "./tanks_picker";
+import { ContentBlock } from '../core/components/content_block';
+import { EncounterPicker } from '../core/components/encounter_picker';
+import { IconPicker } from '../core/components/icon_picker';
+import { SavedDataManager } from '../core/components/saved_data_manager';
+import { SimTab } from '../core/components/sim_tab';
+import * as Tooltips from '../core/constants/tooltips.js';
+import { Encounter } from '../core/encounter';
+import { RaidBuffs } from '../core/proto/common';
+import { SavedEncounter } from '../core/proto/ui';
+import { ActionId } from '../core/proto_utils/action_id';
+import { Raid } from '../core/raid';
+import { EventID } from '../core/typed_event';
+import { AssignmentsPicker } from './components/assignments_picker';
+import { BlessingsPicker } from './components/blessings_picker';
+import { TanksPicker } from './components/tanks_picker';
+import { RaidSimUI } from './raid_sim_ui';
 
 export class SettingsTab extends SimTab {
 	protected simUI: RaidSimUI;
@@ -59,7 +59,7 @@ export class SettingsTab extends SimTab {
 
 	private buildEncounterSettings() {
 		const contentBlock = new ContentBlock(this.column1, 'encounter-settings', {
-			header: { title: 'Encounter' }
+			header: { title: 'Encounter' },
 		});
 
 		new EncounterPicker(contentBlock.bodyElement, this.simUI.sim.encounter, { showExecuteProportion: true }, this.simUI);
@@ -67,7 +67,7 @@ export class SettingsTab extends SimTab {
 
 	private buildConsumesSettings() {
 		const contentBlock = new ContentBlock(this.column1, 'consumes-settings', {
-			header: { title: 'Consumables' }
+			header: { title: 'Consumables' },
 		});
 
 		const container = document.createElement('div');
@@ -88,7 +88,6 @@ export class SettingsTab extends SimTab {
 		// const contentBlock = new ContentBlock(this.column1, 'other-settings', {
 		//   header: {title: 'Other'}
 		// });
-
 		// new BooleanPicker(contentBlock.bodyElement, this.simUI.sim.raid, {
 		// 	label: 'Stagger Stormstrikes',
 		// 	labelTooltip: 'When there are multiple Enhancement Shaman in the raid, causes them to coordinate their Stormstrike casts for optimal SS charge usage.',
@@ -102,7 +101,7 @@ export class SettingsTab extends SimTab {
 
 	private buildTankSettings() {
 		const contentBlock = new ContentBlock(this.column2, 'tanks-settings', {
-			header: { title: 'Tanks' }
+			header: { title: 'Tanks' },
 		});
 
 		new TanksPicker(contentBlock.bodyElement, this.simUI);
@@ -110,7 +109,7 @@ export class SettingsTab extends SimTab {
 
 	private buildAssignmentSettings() {
 		const contentBlock = new ContentBlock(this.column2, 'assignments-settings', {
-			header: { title: 'External Buffs' }
+			header: { title: 'External Buffs' },
 		});
 
 		new AssignmentsPicker(contentBlock.bodyElement, this.simUI);
@@ -118,7 +117,7 @@ export class SettingsTab extends SimTab {
 
 	private buildBlessingsPicker() {
 		const contentBlock = new ContentBlock(this.column3, 'blessings-settings', {
-			header: { title: 'Blessings', tooltip: Tooltips.BLESSINGS_SECTION }
+			header: { title: 'Blessings', tooltip: Tooltips.BLESSINGS_SECTION },
 		});
 
 		this.simUI.blessingsPicker = new BlessingsPicker(contentBlock.bodyElement, this.simUI);
