+++ ui/scss/shared/_modal.scss
@@ -28,3 +28,7 @@
 		}
 	}
 }
+
+.modal-body {
+	gap: var(--bs-modal-padding);
+}
